initial commit - config file
This branch will be for my Keep It Simple Stupid code to connect to
Ellucian Ethos and share with friends. This config file is the start to
have a single place to store configuration variables that are needed in
all of the sample php code. For example, the Ethos API key for this app.