fixes issue #10
The Banner API updated some months ago in order to become more
compliant with standard Ethos API and on par with Colleague API with
regard to searching for person resources. This code update fixes the
previous problem and resolved the problem as described in issue #10 and
gets everything working again.